Revert "Merge pull request #1066 from RumovZ/editor-save-dec"
This reverts commit 994081be34fc61479b740dd9ffd4f7f7be03d8ed, reversing
changes made to 091ea676903b8a723a9da7af408d999efaa0c93a.

@RumoVZ this broke a bunch of operations like 'select notes' and
'set due date'. When the triggered signal is connected to a function,
PyQt looks at the function signature to decide what arguments to pass
it. The wrapper was using *args, so PyQt passes in an extra argument,
which the underlying function didn't expect.

I tried settting __signature__ on the wrapper, but PyQT seems to
ignore it, so we may either need to check all of the existing calls
and add the ignored extra arguments, or create a separate wrapper for
such cases.