Revert "Merge pull request #1066 from RumovZ/editor-save-dec"
This reverts commit a0c47243b648a42f0f1a4a2c10c8b8495a0cfee6, reversing
changes made to 0ab87b733958138ae726b7c7a0c46647ca05edb6.

@RumoVZ this broke a bunch of operations like 'select notes' and
'set due date'. When the triggered signal is connected to a function,
PyQt looks at the function signature to decide what arguments to pass
it. The wrapper was using *args, so PyQt passes in an extra argument,
which the underlying function didn't expect.

I tried settting __signature__ on the wrapper, but PyQT seems to
ignore it, so we may either need to check all of the existing calls
and add the ignored extra arguments, or create a separate wrapper for
such cases.