NF: add an extra option to search card to return one by note
Before moving search to rust, there was an add-on allowing to show a single card by note. It became hard to do it
efficiently as I could not easily change the search anymore, nor get both card id and note id together. Instead, I had
to loop and get all note id of the card id returned by search. For 100 000 cards, the time taken was noticeable.

This feature would not change anything to most user and would ensure that the add-on can be efficient again.

Of course, if this feature is added natively, it would be useful too, since when you want to go over all fields and
tags in browser, duplication of cards is useless