Remove outdated info about VS Code
Pylance is now the default for Python on VS Code, so it should no longer be necessary to suggest switching to it.

I haven't used PyCharm so maybe there is something I'm missing, but for the record I haven't had any issues with Pylance understanding the type annotations. Maybe it's worth checking if this is still true? It's pretty new so it may have had improvements since this was written.