disable automatic mypy invocation in VS Code
The default settings don't understand our folder layout (leading to
errors), and place cache files in the root folder. While there is a
setting to provide extra arguments to mypy which we could use to address
the former, paths like ~/.cache do not appear to be supported. Easier
to just rely on pyright for in-IDE linting, and invoke mypy either
via bazel, or via tools/mypy-watch

+ remove unused dist/ exclude